move testing executors to batch crate
