+++ src/batch/src/lib.rs
@@ -45,3 +45,5 @@ pub mod worker_manager;
 extern crate tracing;
 #[macro_use]
 extern crate risingwave_common;
+
+extern crate self as risingwave_batch;
